requirements.txt
flask==3.1.0
anthropic==0.49.0
gunicorn==23.0.0
python-docx==1.1.2
